requirements.txt
tensorflow==2.15.0
Flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
opencv-python==4.8.0.76
efficientnet==1.1.1
tqdm==4.66.1
pandas==2.0.3
seaborn==0.12.2
requests==2.31.0